requirements.txt
asttokens==2.2.1
backcall==0.2.0
camelot-py==0.9.0
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
fonttools==4.39.3
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
parso==0.8.3
pdfminer.six==20221105
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyparsing==3.0.9
PyPDF2==2.12.1
PySimpleGUI==4.60.4
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.0.2
six==1.16.0
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
tzdata==2023.3
wcwidth==0.2.6
XlsxWriter==3.1.0